Create README.md
ALL PROGRESS SO FAR
I have had many ideas for this project. However, from the start, I gravitated towards something like a game. Most likely, finding a game online and making it better. The first Idea that came to my head was the google dinosaur game. I intended to use this code which I found on the internet, but it turned out to be 2700 lines. As a result, I though they even trying to deal with that code would be too ambitious within itself. After this, I decided to go for a rendition of the Lollipop Ninja game. However, I simply could not figure out how I could change the game in a way to make it mine. I decided to finally go with a code I found on github,com, which is flappy bird. After changing the sprites many times and finally finding a theme that works well, I have decided to make that the project for now. The code can be located at https://github.com/Mariii123/Flappy-bird. Currently. I changed the game into a game where you are driving a ferrari from an overhead view, trying to fit in between two hummers, and you must not hit them or the game will be over. However, all I have truly done is change the sprites around. My goals for the project are to eventually add manybe some moving hummers after maybe 50 points, or to maybe implement some aspects from Lollipop ninja. I would be copying Lollipop ninja by potentially adding an invisibility power-up. Not sure.